--- .travis.yml
@@ -1,10 +1,0 @@
-language: rust
-cache: cargo
-rust:
-  - stable
-  - beta
-  - nightly
-jobs:
-  allow_failures:
-    - rust: nightly
-  fast_finish: true
